Micro-optimize output_prime() a bit more by always copying 24 bytes.
It's safe to copy more bytes than we need so long as we don't exceed
the buffer boundaries, and memcpy() is significantly faster when the
size is fixed, even if we end up copying more bytes than strictly
necessary.

Note that using a temporary buffer is now required because the source
and destination region overlap (where previously, they didn't).

This reduces time spent on printing by about 25% though the total time
decreased only from 14.4 to 13.8 seconds (around 12 seconds calculating,
and 2.4 vs 1.8 seconds printing).